Initial project
These are the files created for my project. The final1 file should be opened using the program Dia (at least that worked for me) and the first query from simple.java does not fully work as intended. Everything else in the program should work though. All of the information was fabricated and is in little to no way factual.